requirements/main.txt
einops
flask
flask_cors
librosa
local_attention
matplotlib
numpy
onnx
# onnxruntime
onnxruntime-gpu   # for gpu
# onnxsim
PyYAML
pyworld
resampy
scikit_learn
scipy
SoundFile
tensorboard
torchcrepe
torchfcpe
tqdm
transformers
protobuf<=4.25.3    # wait to update tensorboard
# pysimplegui==4.60.5
freesimplegui
sounddevice
gradio
pyannote.audio
accelerate
huggingface_hub[torch]
